Refactor to remove lodash and add folder structure
