--- crawler.js
@@ -1,6 +1,0 @@
-import {downloadFromURL} from "./http-download.js";
-import {extractInformation} from "./parser.js";
-
-const seedURL = "https://info.pagnis.in/blog/2020/10/13/basic-financial-hygiene/"
-
-const linksToVisit = [];
