--- sample.html
@@ -1,20 +1,0 @@
-<html>
-<head>
-    <title>HTML Test</title>
-</head>
-<body>
-<header>Test</header>
-<main>
-    <h1>Fancy HTML Page</h1>
-
-    <section>
-        <h2>Links</h2>
-        <div>
-            <a href="https://www.100xengineers.com/">100x Engineering</a>
-            <a href="https://www.google.com/">World's Best Search Engine</a>
-        </div>
-    </section>
-</main>
-<footer></footer>
-</body>
-</html>
