+++ src/http-download.js
@@ -1,11 +1,11 @@
-import fetch from "node-fetch";
+import fetch from 'node-fetch';
 
 const downloadFromURL = async (url) => {
   const response = await fetch(url);
   const body = await response.text();
   const headers = await response.headers;
 
-  return {body, headers};
+  return { body, headers };
 };
 
-export {downloadFromURL};
+export { downloadFromURL };
