+++ src/utils.js
@@ -1,0 +1,1 @@
+export const createLink = (url) => new URL(url);
